+++ src/App.jsx
@@ -1,0 +1,12 @@
+import logo from './logo.svg';
+import './App.css';
+
+function App() {
+  return (
+    <div className="App">
+      <h1>RESTy</h1>
+    </div>
+  );
+}
+
+export default App;
